# optimized.js ~30% faster using FNV Hashing, JS Map, and toString().