Results measured on:
- MacBook Pro 2018 (2.6GHz Intel Core i7 w/16GB)
- Node v15.12.0

```
> benchmark.py # Updated to compare optimized.js original (left) and new (right)
Language      | Optimized Original | Optimized | Notes
------------- | ------------------ | --------- | -----
JavaScript    |               0.86 |      0.59 | by Dani Biro and Flo Hinze
```

I'm sure there's a few more optimizations to be had if someone inspected at the TF (JIT) generated code.