@benhoyt 2021-03-19
Thanks for your contribution! I have allowed a custom hash in the C (and now C++) versions, but I'm loathe to allow it in the other languages. How much of the speedup is to the custom hash, and how much to the other optimizations?