@benhoyt 2021-03-19
Yeah, I think I'll skip it for now, but thanks -- it's good code! I may still back out the C++ custom hash (you do kinda need it in C because it doesn't have hash tables in the stdlib).