@peterwmwong 2021-03-19
👍 

Just to be clear, maybe splitting hairs depending on your definition of a builtin hash table in a language... JS (like C) has no built-in hash/hash table (python `__hash__`, Java `hashCode`, etc). Custom hashing in JS isn't a thing as there is no builtin hashing mechanism :)

The JS spec only provides implementation guidance as :clears throat: "mechanisms that, on average, provide access times that are sublinear on the number of elements in the collection".  As per spec, JS `Map`'s lookup don't "hash" (builtin or custom) a key, but iterate through all contained entries test for equality https://tc39.es/ecma262/#sec-map.prototype.set

Maybe one day...