@benhoyt 2021-03-19
Oh, yeah, I guess I considered plain old `Object` a hash table. It looks like it's a bit more complicated than that in most JS implementations, for example V8 uses optimized structures when small or when possible, but hash table otherwise: http://www.jayconrod.com/posts/52/a-tour-of-v8-object-representation