@carstenschwede 2021-03-19
@peterwmwong raises an important point. `{}` has getters and setters which may look like giving you access to keys in a hash table. But these are not actual keys but object attributes of an `Object` instance, which means missing 'keys' are looked up in the prototype chain and some 'keys' are protected.

So `{}` can't be a hash table because it
- a) can't store some keys (protected attributes, e.g. `__proto__`)
- b) is not empty on startup (e.g. `constructor`)

The current JS implementation is thus buggy and will either fail or produce inaccurate counts if the input text contains the above examples.

This PR should be merged - JavaScript has no native hash tables.